Feature: Add header and file submission form in landing page
The app is correctly initialized. It has a header, a paragraph with a short description of the app's purpose, and a file submission form that allows users to submit up to 25MB of .png, .jpg and .tiff files. CSS is used for most of the styles (most of the ones from Streamlit have been overriden). Next is to keep playing around with the styles, make the footer, and test the form (at least concerning the landing page).